requirements.txt
azure-cosmos
azure-identity
python-dotenv